Non-Integer parameters of simulation not in resulting csv-file #457
parameters of the simulation are only added to the resulting csv
file when they can be parsed into an Integer.

Since other types might be used as well to parameterize a simulation,
they should also be parsed.

We would like to change the parameters of simulation such that
they are just remembered -- instead of being parsed into ints, and
then are written into the CSV file as strings.

This is still work in progress
A series of test simulation have been added here. From there solutions can be easily written.

More Simulations added. I am proposing to handle case of when strings to be handled
using by having random number generator to correspond to it. This is because some
other stats test depend on this int that is parsed in prior implementation

intended to fix #446