Non-Integer parameters of simulation not in resulting csv-file #457
parameters of the simulation are only added to the resulting csv
file when they can be parsed into an Integer.

Since other types might be used as well to parameterize a simulation,
they should also be parsed.

We would like to change the parameters of simulation such that
they are just remembered -- instead of being parsed into ints, and
then are written into the CSV file as strings.

A series of test simulation have been added here. From there solutions can be easily written.

To do : improve the test
fixes  #446