Set stderr/stdout to non-blocking
This makes testing in contexts where 'go test' is writing through a pipe more
reliable. We have found that if the pipe reader falls behind, the logs are
discarded intead of stopping the entire system resulting in false negative test
results.

For production use, either stderr/stdout should be redirected
to a file, so that they never block, or NewFileLogger should be used.