Revert "Set stderr/stdout to non-blocking"
This reverts commit 49d040ef5f1ea3b8d7ac33a9d2203e80cde3c78e.

We found that changing flags on stdout has the undesired side-effect
of changing them on stdin too. We explored some workarounds for this,
but they are gross and have downsides. Instead let's revert this until
it becomes an issue again, and then figure out what to do. (I have also
seen situations where test output for failing tests is lost, and I suspect
it is from this non-blocking change, so one more reason to nuke it.)