Optionally list the WebSocket-side URL in ServerIdentity
This change makes it possible to indicate to the websocket
client where it should contact a given server. In the past, the
websocket side of all servers had to be on port p+1. With
this change, it is possible to indicate in a roster.toml
file the correct URL to use to contact a given server.

Also: Fixed a problem go vet found.