LRU cache with maximum size
This changes the cache to use a LRU policy with a maximum number of entries. This maximum is for the moment static when instantiating the cache.
It also removes the goroutine that cleans the expired items by a simple check when getting elements as we can simply check the tail expiration.

Fixes #273