New format for TLS CommonNames
The previous format for the CommonName was the string encoding of the public key.
However, the string encodings are not intended to be able to do round-trips
from point->string->point. Instead, we use hex(marshal(point)), with a 'Z'
on the front to indicate in a backwards-compatible way that this is a new style
CommonName.

(Why 'Z'? Because the resulting CommonName must be a valid domain name. 'Z'
is an invalid hex digit, so if a new-style common name is sent to an old
onet implementation, an error will be thrown, which will more or less clearly
indicate the problem.)

Fixes #485.