Adding a tracing framework to onet
As logging has its limits, this PR adds a tracing extension to the onet/log
package.
Instead of having to rewrite all the log.* messages in all packages using
onet, this PR tries to be smart to create spans (short executions) and
traces (a set of spans) out of log.* messages.
Unfortunately it's not possible to have it all automatic, so this PR adds
the need of a new log-method:  to help it understand which
spans belong together.
Currently it uses the , which allows all steps of a
protocol to be bundled together.
There are two implementations of the tracing framework: honeycomb.io and
a simulationTracer for debugging.
I'm quite positive that it would be not too difficult to add other
frameworks.