Roster: Restore `GetID` to same hash as `ID`
Even though `Roster.ID` is deprecated, it is still used a lot. The `Roster.GetID` method produces hashes that are different from `ID`, so that comparisons between `ID` and `GetID` fail.

As `GetID` is currently only used for `Roster.Equal`, while `ID` is used in many other methods, this PR makes `GetID` behave the same as `ID`.