Support Ubuntu 20.04 on platform mininet
install_mininet.sh installs fro 20.04 from git, because
mininet releases are rarely made now, and because changes were
needed in mininet itself in order to correctly install on 20.04.
These changes were proposed by me and merged into mininet recently,
allowing this work to proceed.

start.py has a critical simplification to use the default
controller. This is required for Mininet on 20.04 to work
at all, and does not stop 14.04 from working. (Others to be tested still.)

Also: Support local development/debugging with VirtualBox,
via NAT port forwarding. (These changes work and are useful, but
are not sufficient to run an entire simulation because the
logging of monitors needs to know the real IP address of the first
server.)