Initialize ServerIdentities for local tests
While fixing a missing test in byzcoin, I saw that the ServerIdentities are not fully
initialized for local tests. As the new configuration contract in byzcoin also
checks for Address and Description in the ServerIdentity, they also must be
set for local tests.

Once this PR is merged, the onet version should be increased.